Fix remaining ruff lint errors
- Remove unnecessary int() cast in utils.py number_to_human
- Prefix unused unpacked variable with underscore in test_database_types.py
- Convert isinstance tuple to union syntax for UP038 compat
- Remove deprecated UP038 from ruff ignore list

+++ data_diff/utils.py
@@ -315,7 +315,7 @@ def number_to_human(n):
     n = float(n)
     millidx = max(
         0,
-        min(len(millnames) - 1, int(math.floor(0 if n == 0 else math.log10(abs(n)) / 3))),
+        min(len(millnames) - 1, math.floor(0 if n == 0 else math.log10(abs(n)) / 3)),
     )
 
     return f"{n / 10 ** (3 * millidx):.0f}{millnames[millidx]}"
